refactor: move to auth folder
